created ChatInput
