+++ chat-react2/src/components/Home/Main/ChatInput.scss
@@ -1,0 +1,52 @@
+.main__container {
+  & .message__field {
+    display: flex;
+    position: relative;
+    gap: 10px;
+    width: 90%;
+    margin: 0 auto;
+    padding-top: 10px;
+    border-top: 1px solid #e7e7e7;
+    padding-bottom: 10px;
+
+    & input[type="file"] {
+      display: none;
+    }
+
+    & .custom-file-upload {
+      display: inline-block;
+      padding-top: 8px;
+      padding-left: 0px;
+      cursor: pointer;
+    }
+
+    & .send-btn {
+      border: none;
+      font-family: "Roboto";
+      background-color: dodgerblue;
+      color: white;
+      font-weight: 600;
+      border-radius: 4px;
+      cursor: pointer;
+    }
+
+    & .message__area {
+      padding: 10px;
+      height: 45px;
+      border: 1px solid #d9d9d9;
+      font-size: 16px;
+      font-family: "Roboto";
+      border-radius: 10px;
+
+      &::placeholder {
+        font-size: 16px;
+        font-family: inherit;
+      }
+    }
+  }
+
+  & .message__area {
+    width: 100%;
+    resize: none;
+  }
+}
